solve deletion. deleting the file stops distribution.
So even if we restricted deletion to be signed a malicious application can still
delete the file to stop distributing it.

So we'll let it up to the applications to decide if it should delete the file if
it's not on the list anymore or not. and allow simpler implementation by just
deleting the file